ui: Add captive portal detection feature
This feature is a settings option for a Windows notification when connectivity issues due to a captive portal is detected. Detection of a captive portal is determined by adding a route and sending a GET request to a specified IP address by calling CaptivePortalDetection.TestOutsideConnectivity from the broker process. When a captive portal is detected, a Windows notification will be displayed, prompting the user to turn off their VPN and enter login credentials for the captive portal before re-enabling their VPN.